Make changes to PropertyBook design
1. Converted PropertyBook to a singleton class as the data
inside PropertyBook can be seen as a global variable that
should be maintained by a single entity.

2. Introduced a function to help remove all invalid custom
properties. Invalid custom properties are those that the user
adds in directly into the json file but fail to declare them in
the custom property json file.

3. Introduced a function to check if the user has defined
custom properties with clashing names/prefixes in the json file.